+++ src/pages/SignIn/SignIn.style.ts
@@ -23,47 +23,7 @@ export const S = {
       }
     }
   `,
-  InputLabel: styled.label`
-    font-size: 1.4rem;
-    padding-bottom: 1rem;
-    color: ${(props) => props.theme.colors.gray_500};
-  `,
-  ValidWrapper: styled.div`
-    display: flex;
-    & > span {
-      font-size: 1.2rem;
-      color: red;
-      margin-left: 0.6rem;
-    }
 
-    & > img {
-      transform: translateY(-0.5rem);
-      padding-left: 0.6rem;
-      width: 1.7rem;
-      height: 1.7rem;
-    }
-  `,
-  RememberId: styled.div`
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: start;
-
-    & > p {
-      margin-left: 0.6rem;
-      margin-top: 0.2rem;
-      color: ${(props) => props.theme.colors.gray_500};
-      font-size: 1.4rem;
-    }
-  `,
-  ErrorWrapper: styled.div`
-    height: 3rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-weight: 600;
-    color: red;
-  `,
   FindAccount: styled.div`
     display: flex;
     align-items: center;
